rename the module to ipc_communication
