+++ client/ayon_core/tools/loader/ipc_frontend_control.py
@@ -6,7 +6,7 @@ from typing import Any, Callable
 from qtpy import QtCore
 
 from ayon_core.lib.events import QueuedEventSystem
-from ayon_core.ipc_api import WaitCallback
+from ayon_core.ipc_communication import WaitCallback
 from ayon_core.tools.ipc_utils.utils import execute_in_main_thread
 
 from .abstract import FrontendLoaderController
@@ -14,7 +14,7 @@ from .ui.window import LoaderWindow
 
 if typing.TYPE_CHECKING:
     from ayon_core.tools.ipc_utils.utils import CommunicationInfo
-    from ayon_core.ipc_api import RequestMessage
+    from ayon_core.ipc_communication import RequestMessage
     from ayon_core.tools.common_models import (
         TagItem,
         ProductTypeIconMapping,
